[BE022] payment process

+++ src/main/java/com/tttm/birdfarmshop/Utils/Response/OrderResponse.java
@@ -28,7 +28,7 @@ public class OrderResponse {
 
     private String note;
 
-    private Float orderAmount;
+    private Integer orderAmount;
 
     private Date orderDate;
